+++ helpers.lua
@@ -37,30 +37,27 @@ NS.getPercent = function(speed)
   return (speed / BASE_MOVEMENT_SPEED) * 100
 end
 
-NS.formatSpeed = function(speed, decimals, round)
-  if round then
-    local percent = NS.getPercent(speed)
-    local rounded = NS.round(percent)
-    local zeroDecimalsFormat = "%.0f%%"
-    return sformat(zeroDecimalsFormat, rounded)
-  else
-    local singleDecimalFormat = "%.1f%%"
-    local twoDecimalFormat = "%.2f%%"
-    local threeDecimalFormat = "%.3f%%"
-    local fourDecimalFormat = "%.4f%%"
-    local fiveDecimalFormat = "%.5f%%"
+local zeroDecimalFormat = "%.0f%%"
+local singleDecimalFormat = "%.1f%%"
+local twoDecimalFormat = "%.2f%%"
+local threeDecimalFormat = "%.3f%%"
+local fourDecimalFormat = "%.4f%%"
+local fiveDecimalFormat = "%.5f%%"
 
-    if decimals == 1 then
-      return sformat(singleDecimalFormat, NS.getPercent(speed))
-    elseif decimals == 2 then
-      return sformat(twoDecimalFormat, NS.getPercent(speed))
-    elseif decimals == 3 then
-      return sformat(threeDecimalFormat, NS.getPercent(speed))
-    elseif decimals == 4 then
-      return sformat(fourDecimalFormat, NS.getPercent(speed))
-    elseif decimals == 5 then
-      return sformat(fiveDecimalFormat, NS.getPercent(speed))
-    end
+NS.formatSpeed = function(speed, decimals, isDragonRiding)
+  local value = isDragonRiding and speed or NS.getPercent(speed)
+  if decimals == 0 then
+    return sformat(zeroDecimalFormat, value)
+  elseif decimals == 1 then
+    return sformat(singleDecimalFormat, value)
+  elseif decimals == 2 then
+    return sformat(twoDecimalFormat, value)
+  elseif decimals == 3 then
+    return sformat(threeDecimalFormat, value)
+  elseif decimals == 4 then
+    return sformat(fourDecimalFormat, value)
+  elseif decimals == 5 then
+    return sformat(fiveDecimalFormat, value)
   end
 end
 
@@ -94,8 +91,8 @@ NS.IsFlying = function()
   return IsFlying("player")
 end
 
-NS.UpdateText = function(frame, speed, decimals, round)
-  local txt = NS.formatSpeed(speed, decimals, round)
+NS.UpdateText = function(frame, speed, decimals, round, isDragonRiding)
+  local txt = NS.formatSpeed(speed, decimals, round, isDragonRiding)
   if NS.db.global.showlabel then
     if NS.db.global.labeltext then
       local speedWithLabel = sformat("%s %s", NS.db.global.labeltext, txt)